Renamed injectors to components
